% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{df.fill.na}
\alias{df.fill.na}
\title{Replaces NA values in dataframe for closest price}
\usage{
df.fill.na(df.in)
}
\arguments{
\item{df.in}{DAtaframe to be fixed}
}
\value{
A fixed dataframe.
}
\description{
Helper function for BatchGetSymbols. Replaces NA values and returns fixed dataframe.
}
