% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{fix.ticker.name}
\alias{fix.ticker.name}
\title{Fix name of ticker}
\usage{
fix.ticker.name(ticker.in)
}
\arguments{
\item{ticker.in}{A bad ticker name}
}
\value{
A good ticker name
}
\description{
Removes bad symbols from names of tickers. This is useful for naming files with cache system.
}
