% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSP500Stocks.R
\name{GetSP500Stocks}
\alias{GetSP500Stocks}
\title{Function to download the current components of the SP500 index from Wikipedia}
\usage{
GetSP500Stocks(
  do.cache = TRUE,
  cache.folder = file.path(tempdir(), "BGS_Cache")
)
}
\arguments{
\item{do.cache}{Use cache system? (default = TRUE)}

\item{cache.folder}{Where to save cache files? (default = file.path(tempdir(), 'BGS_Cache') )}
}
\value{
A dataframe that includes a column with the list of tickers of companies that belong to the SP500 index
}
\description{
This function scrapes the stocks that constitute the SP500 index from the wikipedia page at https://en.wikipedia.org/wiki/List_of_S%26P_500_companies.
}
\examples{
\dontrun{
df.SP500 <- GetSP500Stocks()
print(df.SP500$tickers)
}
}
