% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{reshape.wide}
\alias{reshape.wide}
\title{Transforms a dataframe in the long format to a list of dataframes in the wide format}
\usage{
reshape.wide(df.tickers)
}
\arguments{
\item{df.tickers}{Dataframe in the long format}
}
\value{
A list with dataframes in the wide format
}
\description{
Transforms a dataframe in the long format to a list of dataframes in the wide format
}
\examples{

my.f <- system.file( 'extdata/ExampleData.rds', package = 'BatchGetSymbols' )
df.tickers <- readRDS(my.f)
l.wide <- reshape.wide(df.tickers)
l.wide
}
