% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{reshape_wide}
\alias{reshape_wide}
\title{Transforms a dataframe in the long format to a list of dataframes in the wide format}
\usage{
reshape_wide(df_in)
}
\description{
Transforms a dataframe in the long format to a list of dataframes in the wide format
}
