% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_cache.R
\name{yf_cachefolder_get}
\alias{yf_cachefolder_get}
\title{Returns the default folder for caching}
\usage{
yf_cachefolder_get()
}
\value{
a path (string)
}
\description{
By default, yfR uses a temp dir to store files.
}
\examples{
print(yf_cachefolder_get())
}
