% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_convert_to_wide.R
\name{yf_converto_to_wide}
\alias{yf_converto_to_wide}
\title{Transforms a dataframe in the long format to a list of dataframes in the wide format}
\usage{
yf_converto_to_wide(df_in)
}
\arguments{
\item{df_in}{Dataframe in the long format (probably the output of yf_get_data())}
}
\value{
A list with dataframes in the wide format (each element is a different column)
}
\description{
Transforms a dataframe in the long format to a list of dataframes in the wide format
}
\examples{

my_f <- system.file("extdata/example_data_yfR.rds", package = "yfR")
df_tickers <- readRDS(my_f)
l_wide <- yf_converto_to_wide(df_tickers)
l_wide
}
