% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_collection.R
\name{yf_get_collection}
\alias{yf_get_collection}
\title{Downloads a collection of data from Yahoo Finance}
\usage{
yf_get_collection(
  collection,
  first_date = Sys.Date() - 15,
  last_date = Sys.Date(),
  do_cache = TRUE,
  cache_folder = yf_get_default_cache_folder()
)
}
\arguments{
\item{collection}{A collection to fetch data (e.g. "SP500", "IBOV", "FTSE" )}

\item{first_date}{The first date of query (Date or character as YYYY-MM-DD)}

\item{last_date}{The last date of query (Date or character as YYYY-MM-DD)}

\item{do_cache}{Use cache system? (default = TRUE)}

\item{cache_folder}{Where to save cache files? (default = yfR::yf_get_default_cache_folder() )}
}
\value{
A dataframe with financial prices from collection
}
\description{
This function will use a set collection of YF data, such as index components and
will download all data from Yahoo Finance using \code{\link{yf_get_data}}.
}
\examples{
\dontrun{
df_yf <- yf_get_collection("IBOV")
}

}
