% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_cache.R
\name{yf_get_default_cache_folder}
\alias{yf_get_default_cache_folder}
\title{Returns the default folder for caching}
\usage{
yf_get_default_cache_folder()
}
\value{
a path (string)
}
\description{
By default, yfR uses a temp dir to store files.
}
\examples{
print(yf_get_default_cache_folder())
}
