% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_dividends.R
\name{yf_get_dividends}
\alias{yf_get_dividends}
\title{Get Yahoo Finance Dividends from a single stock}
\usage{
yf_get_dividends(ticker, first_date, last_date)
}
\arguments{
\item{ticker}{a single ticker symbol}

\item{first_date}{The first date of query (Date or character as YYYY-MM-DD)}

\item{last_date}{The last date of query (Date or character as YYYY-MM-DD)}
}
\value{
a tibble with dividends
}
\description{
This function will use the json api to retrieve dividends from Yahoo finance.
}
\examples{
yf_get_dividends(ticker = "PETR4.SA", first_date = "2010-01-01", last_date = Sys.Date())

}
