% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_index_components.R
\name{yf_get_ibov_stocks}
\alias{yf_get_ibov_stocks}
\title{Function to download the current components of the Ibovespa index from B3 website}
\usage{
yf_get_ibov_stocks(
  do_cache = TRUE,
  cache_folder = yf_get_default_cache_folder(),
  max_tries = 10
)
}
\description{
Function to download the current components of the Ibovespa index from B3 website
}
