% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_single_ticker.R
\name{yf_get_single_ticker}
\alias{yf_get_single_ticker}
\title{An improved version of function quantmod::getSymbols}
\usage{
yf_get_single_ticker(
  ticker,
  i_ticker,
  length_tickers,
  src = "yahoo",
  first_date,
  last_date,
  do_cache = TRUE,
  cache_folder = yf_get_default_cache_folder(),
  df_bench = NULL,
  be_quiet = FALSE,
  thresh_bad_data
)
}
\description{
An improved version of function quantmod::getSymbols
}
