% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_index_components.R
\name{yf_get_sp500_stocks}
\alias{yf_get_sp500_stocks}
\title{Function to download the current components of the SP500 index from Wikipedia}
\usage{
yf_get_sp500_stocks(
  do_cache = TRUE,
  cache_folder = yf_get_default_cache_folder()
)
}
\description{
Function to download the current components of the SP500 index from Wikipedia
}
