% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_index_components.R
\name{yf_index_composition}
\alias{yf_index_composition}
\title{Get current composition of stock indices}
\usage{
yf_index_composition(
  mkt_index,
  do_cache = TRUE,
  cache_folder = yf_cachefolder_get(),
  force_fallback = FALSE
)
}
\arguments{
\item{mkt_index}{the index (e.g. IBOV, SP500, FTSE)}

\item{do_cache}{Use cache system? (default = TRUE)}

\item{cache_folder}{Where to save cache files?
(default = yfR::yf_cachefolder_get() )}

\item{force_fallback}{Logical (TRUE/FALSE). Forces the function to use the
fallback system}
}
\value{
A dataframe with the index compositino (column might vary)
}
\description{
Get current composition of stock indices
}
\examples{
df_sp500 <- yf_index_composition("SP500")
}
