% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_index_components.R
\name{yf_index_list}
\alias{yf_index_list}
\title{Get available indices in package}
\usage{
yf_index_list(print_description = FALSE)
}
\arguments{
\item{print_description}{Logical (TRUE/FALSE) - flag for printing description of
available indices/collections}
}
\value{
A vector of mkt indices
}
\description{
This function will return all available market indices that are registered
in the package.
}
\examples{

indices <- yf_index_list()
indices
}
