% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{calc.ret}
\alias{calc.ret}
\title{Function to calculate returns from a price and ticker vector}
\usage{
calc.ret(P, tickers = rep("ticker", length(P)), type.return = "arit")
}
\description{
Function to calculate returns from a price and ticker vector
}
