% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{reshape.wide}
\alias{reshape.wide}
\title{Transforms a dataframe in the long format to a list of dataframes in the wide format}
\usage{
reshape.wide(df.tickers)
}
\description{
Transforms a dataframe in the long format to a list of dataframes in the wide format
}
