% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_collection.R
\name{yf_get_available_collections}
\alias{yf_get_available_collections}
\title{Returns available collections}
\usage{
yf_get_available_collections()
}
\value{
A string vector with available collections
}
\description{
Returns available collections
}
\examples{
yf_get_available_collections()
}
