% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_collections.R
\name{yf_get_available_collections}
\alias{yf_get_available_collections}
\title{Returns available collections}
\usage{
yf_get_available_collections(print_description = FALSE)
}
\arguments{
\item{print_description}{Logical (TRUE/FALSE) - flag for printing description of
available indices/collections}
}
\value{
A string vector with available collections
}
\description{
Returns available collections
}
\examples{

print(yf_get_available_collections())
}
