% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_index_components.R
\name{yf_get_available_indices}
\alias{yf_get_available_indices}
\title{Get available indices in package}
\usage{
yf_get_available_indices(print_description = FALSE)
}
\arguments{
\item{print_description}{Logical (TRUE/FALSE) - flag for printing description of
available collections}
}
\value{
A vector of mkt indices
}
\description{
This function will return all available market indices that are registered
in the package.
}
\examples{

indices <- yf_get_available_indices()
indices
}
