% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_utils.R
\name{yf_get_default_cache_folder}
\alias{yf_get_default_cache_folder}
\title{Returns the default folder for caching}
\usage{
yf_get_default_cache_folder()
}
\value{
path of default cache folder
}
\description{
Returns the default folder for caching
}
\examples{
yf_get_default_cache_folder
}
