% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_get_index_components.R
\name{yf_get_index_comp}
\alias{yf_get_index_comp}
\title{Get current composition of stock indices}
\usage{
yf_get_index_comp(
  mkt_index,
  do_cache = TRUE,
  cache_folder = yf_get_default_cache_folder()
)
}
\arguments{
\item{mkt_index}{the index (e.g. IBOV, SP500, FTSE)}

\item{do_cache}{Use cache system? (default = TRUE)}

\item{cache_folder}{Where to save cache files? (default = yfR::yf_get_default_cache_folder() )}
}
\value{
A dataframe with the index compositino (column might vary)
}
\description{
Get current composition of stock indices
}
\examples{
df_sp500 <- yf_get_index_comp("SP500")
}
