% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf_live_price.R
\name{yf_live_prices}
\alias{yf_live_prices}
\title{Yahoo Finance Live Prices}
\usage{
yf_live_prices(.ticker)
}
\arguments{
\item{.ticker}{Ticker de interesse}
}
\value{
Tibble
}
\description{
Yahoo Finance Live Prices
}
\examples{
yfR::yf_live_prices("PETR4.SA")

}
